+++ playtime-0.2.0/src/controllers/accounts-controller.js
@@ -1,3 +1,4 @@
+/* eslint-disable func-names */
 import { db } from "../models/db.js";
 
 export const accountsController = {