+++ playtime-0.4.0/src/models/json/playlist-json-store.js
@@ -1,6 +1,6 @@
 import { v4 } from "uuid";
-// eslint-disable-next-line import/no-unresolved
-import { JSONFile, Low } from "lowdb";
+import { Low } from "lowdb";
+import { JSONFile } from "lowdb/node";
 import { trackJsonStore } from "./track-json-store.js";
 
 const db = new Low(new JSONFile("./src/models/json/playlists.json"));